pyproject.toml
[project]
name = "podcaster"
version = "0.1.0"
description = "Convert epub chapters to narrated MP3 using Gemini TTS"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.0",
    "pydub>=0.25",
    "ebooklib>=0.18",
    "beautifulsoup4>=4.12",
    "tejas-config",
]

[project.scripts]
podcaster = "main:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["epub", "tts", "main", "db"]

[tool.uv.sources]
tejas-config = { path = "../tejas-config" }
